requirements.txt
alembic
aiogram
fastapi
pydantic
pydantic-settings
sqlalchemy
python-jose
psycopg2
bcrypt
uvicorn
environs
email-validator
redis